@@ -36,7 +36,7 @@ public class LongTypeConverter extends NumberTypeConverterSupport implements Typ
         Number number = parse(input, errors);
         Long retval = null;
         if (errors.size() == 0) {
-            retval = new Long(number.longValue());
+            retval = Long.valueOf(number.longValue());
         }
 
         return retval;
