@@ -21,16 +21,16 @@ import java.util.Collection;
 /**
  * This is an enumeration of all of the valid HTTP request method types
  * which can be used for event handling within RestActionBean classes.
- * 
+ *
  * @author Rick Grashel
  */
 public enum HttpRequestMethod
 {
-    GET, POST, HEAD, PUT, DELETE;
+    GET, POST, HEAD, PUT, DELETE, OPTIONS, TRACE, CONNECT, PATCH;
 
     /**
      * Returns a collection of all of the values for this enumeration.
-     * 
+     *
      * @return Collection of all values for this enumeration
      */
     public static Collection<HttpRequestMethod> all() {
