@@ -19,10 +19,13 @@ import java.util.List;
 import java.util.Map;
 import java.util.SortedMap;
 import java.util.TreeMap;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
 import javax.servlet.Filter;
 
 import net.sourceforge.stripes.action.ActionBean;
+import net.sourceforge.stripes.action.Message;
 import net.sourceforge.stripes.controller.ActionResolver;
 import net.sourceforge.stripes.controller.AnnotatedClassActionResolver;
 import net.sourceforge.stripes.controller.StripesConstants;
@@ -98,7 +101,7 @@ public class MockRoundtrip {
 
     /**
      * Constructor that will create a request suitable for the provided servlet context and
-     * URL. Note that in general the contructors that take an ActionBean Class object are preferred
+     * URL. Note that in general the constructors that take an ActionBean Class object are preferred
      * over those that take a URL.  Constructs a new session for the request.
      *
      * @param context the MockServletContext that will receive this request
@@ -284,6 +287,23 @@ public class MockRoundtrip {
     public ValidationErrors getValidationErrors() {
         ActionBean bean = (ActionBean) this.request.getAttribute(StripesConstants.REQ_ATTR_ACTION_BEAN);
         return bean.getContext().getValidationErrors();
+    }
+
+    /**
+     * Gets the {@link List} of {@link Message}s that were produced by the request.
+     * This should be used instead of obtaining the messages from the
+     * {@link net.sourceforge.stripes.action.ActionBeanContext} as the context is bound to the
+     * {@link net.sourceforge.stripes.controller.FlashScope}.
+     *
+     * @return
+     */
+    public List<Message> getMessages() {
+        Object attribute = this.request.getAttribute(StripesConstants.REQ_ATTR_MESSAGES);
+        if (attribute == null) {
+            return null;
+        }
+
+        return (List<Message>) attribute;
     }
 
     /**
