@@ -18,7 +18,7 @@ package net.sourceforge.stripes.tag;
  * <p>This default implementation of the TagErrorRenderer interface sets the html class
  * attribute to 'error'.  More specifically, if the tag had no previous CSS class, it
  * will have its class attribute set to error. If it previously had a CSS class attribute,
- * e.g. class="foo", then it's class attribute will be re-written as class="foo error",
+ * e.g. class="foo", then its class attribute will be re-written as class="foo error",
  * which instructs the browser to apply both styles, with error taking precedence. The
  * use of a single class name allows applications to define a single style for all input
  * fields, and then override it for specific fields as they choose.</p>
@@ -65,7 +65,7 @@ public class DefaultTagErrorRenderer implements TagErrorRenderer {
     }
 
     /**
-     * Resets the tag's class attribute to it's original value in case the tag gets pooled.
+     * Resets the tag's class attribute to its original value in case the tag gets pooled.
      */
     public void doAfterEndTag() {
         tag.setCssClass(oldCssClass);
